moved pages from components folder to src folder, and added tests folder.
